+++ library/src/main/java/com/wultra/android/mtokensdk/operation/IOperationsService.kt
@@ -103,10 +103,11 @@ interface IOperationsService {
      * If operations are already polling, this call is ignored
      * and the polling interval won't be changed.
      *
-     * @param pollingInterval Polling interval in milliseconds
-     * @param delayStart When true, polling starts after the first [pollingInterval] time passes
+     * @param pollingInterval Polling interval in milliseconds, default value is 7s and minimum is 5s
+     * @param delayStart When true, polling starts after the first [pollingInterval] passes
+     *                   - By default it is set to false and polling starts immediately.
      */
-    fun startPollingOperations(pollingInterval: Long, delayStart: Boolean)
+    fun startPollingOperations(pollingInterval: Long = 7_000, delayStart: Boolean = false)
 
     /**
      * Stops operation polling