+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 - Removed currentServerTime() method [(#139)](https://github.com/wultra/mtoken-sdk-android/pull/139)
 - Improved operations handling [(#138)](https://github.com/wultra/mtoken-sdk-android/pull/138)
 - Added Huawei notification support [(#136)](https://github.com/wultra/mtoken-sdk-android/pull/136)
+- Implement default and minimum pollingInterval [(#142)](https://github.com/wultra/mtoken-sdk-android/pull/142)
 
 ## 1.9.0 (Jan 25, 2024)